Continuar bosquejando...
Creo que ahora tengo una idea más o menos clara de por dónde comenzar a trabajar...

nota: Partía con ventaja porque ya he realizado anteriormente muchas aplicaciones con albaranes. (Y además esto es solo un ejercicio que voy modelando según me parece :-)
Si llega a ser un dominio completamente nuevo para mi, el enfoque seria distinto: comenzar por implementar una funcionalidad pequeña, luego otra, luego otra,... según las fuera pidiendo el cliente ; refactorizando cuando se viera necesario; con mucho diálogo, afianzando poco a poco un "ubiquitous language";  hasta que se fuera perfilando poco a poco una arquitectura, a medida que el conocimiento del dominio fuera cada vez mayor...

+++ src/main/java/es/susosise/albaranes/albaranes/AlbaranLinea.java
@@ -1,11 +1,12 @@
 package es.susosise.albaranes.albaranes;
 
-import es.susosise.albaranes.articulos.Articulo;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+
+import es.susosise.albaranes.articulos.Articulo;
 
 @Entity
 @Table(name = "albaranes_lineas")
@@ -15,6 +16,6 @@ public class AlbaranLinea {
     private Long idInterno;
 
     Articulo articulo;
-    Double cantidad;
+    Integer cantidad;
 
 }
